requirements.txt
HTMLParser==0.0.2
Pillow==3.1.0
argparse==1.4.0
protobuf==3.0.0-alpha-1
pycrypto==2.6.1
python-axolotl==0.1.7
python-axolotl-curve25519==0.1
python-dateutil==2.4.2
six==1.10.0
yowsup2==2.4.48
